apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
supabase==2.0.3
httpx>=0.24.0,<0.25.0
pillow==10.1.0
numpy==1.26.2
torch>=2.0.0
torchvision>=0.18.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
# Optional but recommended for Qwen 3.5 Vision (faster load + inference):
# pip install unsloth

